% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf.cal.meta.R
\name{bf.cal.meta}
\alias{bf.cal.meta}
\title{Bayes Factor Calculation Scheme for META prior}
\usage{
bf.cal.meta(data, hyperparam)
}
\arguments{
\item{data}{A dataset which is constructed by pairs of coefficient
values \eqn{ \beta } and standard errors \eqn{ se(\beta)}.}

\item{hyperparam}{A two-dimensional vector denoting all the grid points,
namely, \eqn{\phi} x \eqn{\omega}.}
}
\value{
A list records all the log scale bayes factor values.
}
\description{
A function that calculates bayes factor for each data pair on each grid point
 in log scale.
}
