% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf.loglik.R
\name{bf.loglik}
\alias{bf.loglik}
\title{Bayes Factor Loglikelihood Function}
\usage{
bf.loglik(w, bf)
}
\arguments{
\item{w}{The current weight vector}

\item{bf}{A vector recording all the bayes factor values.}
}
\value{
Negative summation of loglikelihood values.
}
\description{
Calculate the updated loglikelihood value in EM algorithm, and to evaluate whether converge or not.
}
