% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf.weighted_sum.R
\name{bf.weighted_sum}
\alias{bf.weighted_sum}
\title{Bayes Factor Weighted Summation}
\usage{
bf.weighted_sum(w, bf, i)
}
\arguments{
\item{w}{Input weight vector.}

\item{bf}{Input bayes factor vector}

\item{i}{Individual index.}
}
\value{
Weighted sum for bayes factor in log scale.
}
\description{
A function calculates the weighted summation of bayes factor.
}
