% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cefee.R
\name{cefee}
\alias{cefee}
\title{Barbeta Integration Function}
\usage{
cefee(x, param = data, z = i, size = m)
}
\arguments{
\item{x}{Unknown parameter, index the variable to be integrated out.(\eqn{beta})}

\item{param}{Parameter list.}

\item{z}{Index for the unit.(i)}

\item{size}{Index for the number of replicates. (m)}
}
\value{
The value for the function.
}
\description{
Barbeta Integration Function
}
