% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetero.lfdr.R
\name{hetero.lfdr}
\alias{hetero.lfdr}
\title{Local False Discovery Rate Evaluation}
\usage{
hetero.lfdr(cat, fdr.level)
}
\arguments{
\item{fdr.level}{Rejection level for Local false discovery rate, if not specified, the rejection
decision procedure won't be run.}

\item{catfinal}{The final individual-level probabilities of falling into three
categories, separately.}
}
\value{
A list that preserves local false discovery rate and 5tthe corresponding reject
 decision if called.
}
\description{
Local False Discovery Rate Evaluation
}
