% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetero.R
\name{hetero}
\alias{hetero}
\title{Heterogeneity Evaluation}
\usage{
hetero(data, krep = NULL, kirep = NULL, oa = NULL,
  sq_em_tol = 1e-07)
}
\arguments{
\item{data}{A dataset which is constructed by pairs of coefficient
values \eqn{ \beta } and standard errors \eqn{ se(\beta)}.}

\item{krep}{A vector, denoting all the k values constructing the reproducible signals. If not specified,
the default one is c(0.01,0.462,0.513,0.618), which corresponds to the several prior values satisfy that
\eqn{Pr(\beta_{i,1}, \beta_{i,2} have a same sign)\ge 0.90.}}

\item{kirep}{A vector, denoting all the k values constructing the irreproducible signals.If not specified,
the default one is c(2.0,10.0), which corresponds to the several prior values satisfy that
\eqn{Pr(\beta_{i,1}, \beta_{i,2} have a same sign)=0.57,0.5.}}

\item{oa}{A vector, denoting all the \eqn{\omega} values to be set on the grid. If not specified, \eqn{\omega} grid will
be generated by a dynamic procedure, which is determined by the data structure, namely,
\eqn{min(\omega)=min(se(\beta)/2)}, \eqn{max(\omega)=max(\sqrt{\beta^2-se(\beta)^2})}.}

\item{sq_em_tol}{A small, positive scalar that determines when iterations should be terminated in squarem algorithm.
The default value is \eqn{1e-7}.}
}
\value{
A list with the following components:
\describe{
\item{gridweight}{ The final optimal weight vector evaluated on each grid point.}
\item{ind_prob}{ A matrix denoting the converged probability for each individual being inside the
three different groups, namely, the null, the reproducible and the irreproducible group.}
\item{est_prop}{ The estimated proportion value for the three different groups, namely, the null,
the reproducible and the irreproducible group.}
}
}
\description{
Evaluating the overall and individually heterogeneity and reproducibility
for the given individuals(units) shared in different replicates.
}
\examples{
data("heterodata")
hetero.out<-hetero(heterodata)
names(hetero.out)

}
